Cleaned up code and tested splash screen
